+++ R/power-distribution.R
@@ -3,7 +3,7 @@
 #' Power distribution
 #'
 #' Density, distribution function, quantile function and random generation
-#' for the Pareto distribution.
+#' for the power distribution.
 #'
 #' @param x,q	            vector of quantiles.
 #' @param p	              vector of probabilities.