+++ NEWS.md
@@ -1,7 +1,7 @@
 
 ### 1.7.1
 
-* Added Dirichlet-multinomial distribution
+* Added Dirichlet-multinomial and beta prime distributions
 * Code clean-up for categorical distribution
 * Improvements in documentation and examples (e.g. bivariate Poisson,
   bivariate normal)